Refactored the app so that the model handles all the data related manipulations
